fixe Delete avec la gestion d'erreur

+++ server.js
@@ -16,7 +16,8 @@ else {
         const request_url = req.url;
         const { pathname, query } = url.parse(request_url, true)
 
-
+        /* -------------------------GET------------------------*/
+        
         if (req.method == 'GET') {
             if (pathname === '/') {
                 const { name } = query
@@ -25,6 +26,8 @@ else {
 
             }
         }
+
+        /* -------------------------POST------------------------*/
 
         if (req.method == 'POST') {
             if (pathname === '/students') {
@@ -59,7 +62,8 @@ else {
             }
         }
 
-        // il faut verifie si on met des lettres ou si l'id n'existe pas 
+      
+        /* -------------------------PUT------------------------*/
 
         if (req.method == 'PUT') {
             let putId = pathname.split('/')
@@ -104,34 +108,51 @@ else {
             })
         }
 
-        if (req.method == 'DELETE'){
-            let putId = pathname.split('/')
-            putId = putId[2]
+        /* -------------------------DELETE------------------------*/
+
+        if (req.method == 'DELETE') {
+            let splitPathname = pathname.split('/')
+            let ressource = splitPathname[1]
+            let putId = splitPathname[2]
+            let matchDelete = []
             let json = ""
             let globalJson = []
-            putId = parseInt(putId)
-       
-            if (typeof putId === 'number') {
-                if (fs.existsSync(LOCAL_DATABASE)) {
-                    json = require(`./${LOCAL_DATABASE}`) //on lit et parse en string
-                    for (const key in json) {
-                        if (json[key].id != putId) {
-                            globalJson.push(json[key])
-                            console.log("l'objet avec l'id "+json[key].id+" a été ajouté")
+            let isExist = false
+
+            matchDelete = putId.match(/(^[0-9]+$)/g) // ON VERIFIE SI ON MATCH AVEC UN NOMBRE
+            if (ressource == 'students') { // ON VERIFIE SI LA RESSOURCE EST STUDENTS
+                if (matchDelete != null) { // ON VERIFIE SI IL Y UN NOMBRE
+                    putId = parseInt(matchDelete[0])
+                    if (fs.existsSync(LOCAL_DATABASE)) {
+                        json = require(`./${LOCAL_DATABASE}`) //on lit et parse en string
+                        //ON VERIFIE SI L'ID EXISTE
+                        for (const k in json) {
+                            if (json[k].id == putId) {
+                                isExist = true
+                            }
                         }
+                        // SI IL EXISTE ON VA CHERCHE LA LIGNE A SUPPRIMER
+                        if (isExist == true) {
+                            for (const key in json) {
+                                if (json[key].id != putId) {
+                                    globalJson.push(json[key]) // ON AJOUTE TOUT LES OBJECTS QUI N'ONT PAS L'ID MENTIONNE
+                                    fs.writeFileSync(LOCAL_DATABASE, JSON.stringify(globalJson, null, 4))
+                                    console.log("l'objet avec l'id " + json[key].id + " a été ajouté")
+                                }
+                            }
+                        } else {
+                            console.log("l'id " + putId + " n'existe pas")
+                        }
+                    } else {
+                        console.log(` le fichier ${LOCAL_DATABASE} ne peut pas etre modifié si il n'existe pas`)
                     }
                 } else {
-                    console.log(` le fichier ${LOCAL_DATABASE} ne peut pas etre modifié si il n'existe pas`)
+                    console.log("l'id renseigné n'est pas un nombre ")
                 }
             } else {
-                console.log("l'id renseigné n'est pas un nombre ")
+                console.log("Votre ressource doit être 'students'")
             }
-            console.log(globalJson)
-            fs.writeFileSync(LOCAL_DATABASE, JSON.stringify(globalJson, null, 4))
-
-            
         }
-
         res.end();
 
     }).listen(port)
